Deduplicated header code.
